refactor: extract mra review-protocol + secret-strip into modules (#76)
mra.ts (1163 lines) mixed the pure review contract with the spawn machinery.
Pull the parts that don't spawn out:

- mra-env.ts — strippedChildEnv + the secret-shaped-env denylist, shared by
  every mra spawn path (ask, review, analyze, capability probe).
- mra-review-protocol.ts — MraReviewResult + strategy/provider types,
  buildReviewArgv, parseReviewStdout, reviewEnv. Pure; reviewEnv pulls
  strippedChildEnv from mra-env. Clean DAG, no cycle.

runMraReview keeps the spawn wiring and imports these; both modules are
re-exported from mra.ts so callers and tests import them from adapters/mra
unchanged. mra.ts 1163 → 1038 lines.

Behaviour-preserving: 1009 cli tests pass, typecheck clean. The argv/parse/env
contract already had heavy coverage in mra-review.test.ts (buildReviewArgv 11,
parseReviewStdout 9, reviewEnv 15 assertions), which now exercises the modules.
